# minor Span impl improvements, no locks on atomic assignments