--- a/src/OpenTelemetry/Trace/Span.cs
+++ b/src/OpenTelemetry/Trace/Span.cs
@@ -13,11 +13,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
+
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using OpenTelemetry.Context.Propagation;
+using OpenTelemetry.Internal;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace.Configuration;
 using OpenTelemetry.Trace.Export;
@@ -34,7 +36,7 @@ namespace OpenTelemetry.Trace
         private readonly TracerConfiguration tracerConfiguration;
         private readonly SpanProcessor spanProcessor;
         private readonly DateTimeOffset startTimestamp;
-        private readonly object @lock = new object();
+        
         private EvictingQueue<KeyValuePair<string, object>> attributes;
         private EvictingQueue<Event> events;
         private EvictingQueue<Link> links;
@@ -103,26 +105,9 @@ namespace OpenTelemetry.Trace
         /// <inheritdoc/>
         public Status Status
         {
-            get
-            {
-                lock (this.@lock)
-                {
-                    return this.StatusWithDefault;
-                }
-            }
+            get => this.StatusWithDefault;
 
-            set
-            {
-                if (!this.IsRecordingEvents)
-                {
-                    return;
-                }
-
-                lock (this.@lock)
-                {
-                    this.status = value.IsValid ? value : throw new ArgumentException(nameof(value));
-                }
-            }
+            set => this.status = value.IsValid ? value : throw new ArgumentException(nameof(value));
         }
 
         public ActivitySpanId ParentSpanId => this.Activity.ParentSpanId;
@@ -174,92 +159,84 @@ namespace OpenTelemetry.Trace
         /// <inheritdoc />
         public void UpdateName(string name)
         {
+            if (this.hasEnded)
+            {
+                OpenTelemetrySdkEventSource.Log.UnexpectedCallOnEndedSpan("UpdateName");
+                return;
+            }
+
             this.Name = name ?? throw new ArgumentNullException(nameof(name));
         }
 
         /// <inheritdoc/>
         public void SetAttribute(KeyValuePair<string, object> keyValuePair)
         {
-            if (keyValuePair.Key == null || keyValuePair.Value == null)
+            if (keyValuePair.Key == null)
             {
-                throw new ArgumentNullException(nameof(keyValuePair));
+                throw new ArgumentNullException(nameof(keyValuePair.Key));
             }
 
-            if (!this.IsRecordingEvents || this.hasEnded)
+            if (keyValuePair.Value == null)
+            {
+                throw new ArgumentNullException(nameof(keyValuePair.Value));
+            }
+
+            if (this.hasEnded)
+            {
+                OpenTelemetrySdkEventSource.Log.UnexpectedCallOnEndedSpan("UpdateName");
+                return;
+            }
+
+            if (!this.IsRecordingEvents)
             {
                 return;
             }
 
-            lock (this.@lock)
+            if (this.attributes == null)
             {
-                if (this.attributes == null)
-                {
-                    this.attributes = new EvictingQueue<KeyValuePair<string, object>>(this.tracerConfiguration.MaxNumberOfAttributes);
-                }
-
-                this.attributes.AddEvent(new KeyValuePair<string, object>(keyValuePair.Key, keyValuePair.Value));
+                this.attributes = new EvictingQueue<KeyValuePair<string, object>>(this.tracerConfiguration.MaxNumberOfAttributes);
             }
+
+            this.attributes.Add(new KeyValuePair<string, object>(keyValuePair.Key, keyValuePair.Value));
         }
 
         /// <inheritdoc/>
         public void AddEvent(string name)
         {
-            if (name == null)
+            if (this.hasEnded)
             {
-                throw new ArgumentNullException(nameof(name));
+                OpenTelemetrySdkEventSource.Log.UnexpectedCallOnEndedSpan("AddEvent");
+                return;
             }
 
-            if (!this.IsRecordingEvents || this.hasEnded)
+            if (!this.IsRecordingEvents)
             {
                 return;
             }
 
-            lock (this.@lock)
-            {
-                if (this.events == null)
-                {
-                    this.events =
-                        new EvictingQueue<Event>(this.tracerConfiguration.MaxNumberOfEvents);
-                }
-
-                this.events.AddEvent(new Event(name, PreciseTimestamp.GetUtcNow()));
-            }
+            this.AddEvent(new Event(name, PreciseTimestamp.GetUtcNow()));
         }
 
         /// <inheritdoc/>
         public void AddEvent(string name, IDictionary<string, object> eventAttributes)
         {
-            if (name == null)
-            {
-                throw new ArgumentNullException(nameof(name));
-            }
-
             if (eventAttributes == null)
             {
                 throw new ArgumentNullException(nameof(eventAttributes));
             }
 
-            if (!this.IsRecordingEvents || this.hasEnded)
+            if (this.hasEnded)
+            {
+                OpenTelemetrySdkEventSource.Log.UnexpectedCallOnEndedSpan("AddEvent");
+                return;
+            }
+
+            if (!this.IsRecordingEvents)
             {
                 return;
             }
 
-            lock (this.@lock)
-            {
-                if (this.hasEnded)
-                {
-                    // logger.log(Level.FINE, "Calling AddEvent() on an ended Span.");
-                    return;
-                }
-
-                if (this.events == null)
-                {
-                    this.events =
-                        new EvictingQueue<Event>(this.tracerConfiguration.MaxNumberOfEvents);
-                }
-
-                this.events.AddEvent(new Event(name, PreciseTimestamp.GetUtcNow(), eventAttributes));
-            }
+            this.AddEvent(new Event(name, PreciseTimestamp.GetUtcNow(), eventAttributes));
         }
 
         /// <inheritdoc/>
@@ -270,27 +247,24 @@ namespace OpenTelemetry.Trace
                 throw new ArgumentNullException(nameof(addEvent));
             }
 
+            if (this.hasEnded)
+            {
+                OpenTelemetrySdkEventSource.Log.UnexpectedCallOnEndedSpan("AddEvent");
+                return;
+            }
+
             if (!this.IsRecordingEvents)
             {
                 return;
             }
 
-            lock (this.@lock)
+            if (this.events == null)
             {
-                if (this.hasEnded)
-                {
-                    // logger.log(Level.FINE, "Calling AddEvent() on an ended Span.");
-                    return;
-                }
-
-                if (this.events == null)
-                {
-                    this.events =
-                        new EvictingQueue<Event>(this.tracerConfiguration.MaxNumberOfEvents);
-                }
-
-                this.events.AddEvent(addEvent);
+                this.events =
+                    new EvictingQueue<Event>(this.tracerConfiguration.MaxNumberOfEvents);
             }
+
+            this.events.Add(addEvent);
         }
 
         /// <inheritdoc/>
@@ -301,26 +275,23 @@ namespace OpenTelemetry.Trace
                 throw new ArgumentNullException(nameof(link));
             }
 
+            if (this.hasEnded)
+            {
+                OpenTelemetrySdkEventSource.Log.UnexpectedCallOnEndedSpan("AddLink");
+                return;
+            }
+
             if (!this.IsRecordingEvents)
             {
                 return;
             }
 
-            lock (this.@lock)
+            if (this.links == null)
             {
-                if (this.hasEnded)
-                {
-                    // logger.log(Level.FINE, "Calling addLink() on an ended Span.");
-                    return;
-                }
-
-                if (this.links == null)
-                {
-                    this.links = new EvictingQueue<Link>(this.tracerConfiguration.MaxNumberOfLinks);
-                }
-
-                this.links.AddEvent(link);
+                this.links = new EvictingQueue<Link>(this.tracerConfiguration.MaxNumberOfLinks);
             }
+
+            this.links.Add(link);
         }
 
         /// <inheritdoc/>
@@ -333,6 +304,7 @@ namespace OpenTelemetry.Trace
         {
             if (this.hasEnded)
             {
+                OpenTelemetrySdkEventSource.Log.UnexpectedCallOnEndedSpan("End");
                 return;
             }
 
@@ -365,11 +337,6 @@ namespace OpenTelemetry.Trace
                 throw new ArgumentNullException(nameof(key));
             }
 
-            if (!this.IsRecordingEvents)
-            {
-                return;
-            }
-
             this.SetAttribute(new KeyValuePair<string, object>(key, value));
         }
 
@@ -379,11 +346,6 @@ namespace OpenTelemetry.Trace
             if (key == null)
             {
                 throw new ArgumentNullException(nameof(key));
-            }
-
-            if (!this.IsRecordingEvents)
-            {
-                return;
             }
 
             this.SetAttribute(new KeyValuePair<string, object>(key, value));
@@ -397,11 +359,6 @@ namespace OpenTelemetry.Trace
                 throw new ArgumentNullException(nameof(key));
             }
 
-            if (!this.IsRecordingEvents)
-            {
-                return;
-            }
-
             this.SetAttribute(new KeyValuePair<string, object>(key, value));
         }
 
@@ -411,11 +368,6 @@ namespace OpenTelemetry.Trace
             if (key == null)
             {
                 throw new ArgumentNullException(nameof(key));
-            }
-
-            if (!this.IsRecordingEvents)
-            {
-                return;
             }
 
             this.SetAttribute(new KeyValuePair<string, object>(key, value));