Fixes #271 

**[update] No perf improvements are coming because Span has to be thread-safe, ignore numbers below.**

In the majority of cases, Span is not used concurrently. Even when attributes or events are added, it does not normally happen concurrently.

Let's declare span as not thread-safe and let rare instrumentations that need concurrency do the locking.
If we'll find it's common with e.g. Events, we can create lazy lock when first event is added.

this change removes locking which this simplifies code and reduces allocations 


**After**

|                           Method |        Mean |      Error |     StdDev | Allocated |
|--------------------------------- |------------:|-----------:|-----------:|----------:|
|               CreateSpan_Sampled |   772.27 ns |  7.1575 ns |  6.6951 ns |     560 B |
|    CreateSpan_Attributes_Sampled |   929.65 ns |  8.1989 ns |  7.6692 ns |    1248 B |
|     CreateSpan_Propagate_Sampled | 1,859.91 ns | 19.3340 ns | 17.1391 ns |     933 B |
| CreateSpan_Attributes_NotSampled |   795.95 ns |  7.8778 ns |  7.3689 ns |     560 B |
|                  CreateSpan_Noop |    38.63 ns |  0.3342 ns |  0.2962 ns |         - |
|       CreateSpan_Attributes_Noop |    44.45 ns |  0.3592 ns |  0.3360 ns |         - |
|        CreateSpan_Propagate_Noop |    41.12 ns |  0.4033 ns |  0.3773 ns |         - |

**Before**

|                           Method |        Mean |      Error |     StdDev | Allocated |
|--------------------------------- |------------:|-----------:|-----------:|----------:|
|               CreateSpan_Sampled |   775.89 ns | 10.9065 ns |  9.6683 ns |     592 B |
|    CreateSpan_Attributes_Sampled | 1,016.07 ns |  6.3040 ns |  5.5883 ns |    1280 B |
|     CreateSpan_Propagate_Sampled | 1,894.26 ns | 22.0567 ns | 20.6319 ns |     966 B |
| CreateSpan_Attributes_NotSampled |   802.46 ns |  5.3171 ns |  4.9736 ns |     592 B |
|                  CreateSpan_Noop |    38.56 ns |  0.2433 ns |  0.2275 ns |         - |
|       CreateSpan_Attributes_Noop |    44.65 ns |  0.3581 ns |  0.3350 ns |         - |
|        CreateSpan_Propagate_Noop |    41.41 ns |  0.3090 ns |  0.2891 ns |         - |