@pcwiese 2019-10-21
Do you need to explicitly say that Span is not thread-safe in the class documentation? Should that documentation be on ISpan itself in the API?