@lmolkova 2019-10-21
It seems I sent this PR without reading the spec: 

["All methods of Span are safe to be called concurrently"}(https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/concurrency.md)

So Span must be thread safe 

I'll update this PR to keep some of its goodness and return lock back